+++ packages/plugin-packages/mashroom-portal-default-theme/src/backend/dev_server.ts
@@ -100,7 +100,8 @@ app.get('/', (req: Request, res: Response) => {
         },
         portalResourcesHeader: '',
         portalResourcesFooter: '',
-        portalLayout: `
+        portalLayout: 'deprecated',
+        pageContent: `
             <div class="row">
                 <div class="mashroom-portal-app-area col-md-6" id="app-area1">
                     <div class="mashroom-portal-app-wrapper">