+++ packages/plugin-packages/mashroom-portal/src/backend/utils/render_utils.ts
@@ -6,7 +6,11 @@ import defaultTemplateAppError from '../theme/default_template_app_error';
 import type {Request, Response} from 'express';
 import type {MashroomLogger} from '@mashroom/mashroom/type-definitions';
 import type {MashroomPortalPageRenderModel} from '../../../type-definitions';
-import {MashroomPortalAppErrorRenderModel, MashroomPortalAppWrapperRenderModel} from '../../../type-definitions';
+import {
+    MashroomPortalAppErrorRenderModel,
+    MashroomPortalAppWrapperRenderModel,
+} from '../../../type-definitions';
+import type {MashroomPortalPageAppsInfo} from '../../../type-definitions/internal';
 
 export const renderPage = async (themeExists: boolean, model: MashroomPortalPageRenderModel, req: Request, res: Response, logger: MashroomLogger): Promise<string> => {
     const fallback = () => minimalTemplatePortal(model);
@@ -63,6 +67,52 @@ export const renderAppErrorToClientTemplate = async (themeExists: boolean, messa
     return renderAppError(themeExists, model, req, res, logger);
 }
 
+export const renderPageContent = async (portalLayout: string, portalAppInfo: MashroomPortalPageAppsInfo, themeExists: boolean, messages: (key: string) => string, req: Request, res: Response, logger: MashroomLogger): Promise<string> => {
+    let portalContent = portalLayout;
+    const appAreas = Object.keys(portalAppInfo);
+    const promises: Array<Promise<Array<string>>> = [];
+    for (const appAreaId of appAreas) {
+        promises.push(
+            Promise.all(
+                portalAppInfo[appAreaId].map(({pluginName, appSetup}) => {
+
+                    // TODO: check if app supports SSR
+                    const appSSRHtml = null;
+
+                    const model: MashroomPortalAppWrapperRenderModel = {
+                        appId: appSetup.appId,
+                        pluginName,
+                        safePluginName: getSafePluginName(pluginName),
+                        title: appSetup.title || pluginName,
+                        messages,
+                        appSSRHtml,
+                    };
+                    return renderAppWrapper(themeExists, model, req, res, logger);
+                })
+            ).then(
+                (result) => result,
+                (error) => {
+                    logger.error(`Rendering apps in appArea ${appAreaId} failed`, error);
+                    return [];
+                }
+            )
+        );
+    }
+
+    const appAreasAppHtmls = await Promise.all(promises);
+    appAreas.forEach((appAreaId, idx) => {
+       const appHtmls = appAreasAppHtmls[idx];
+       if (appHtmls.length > 0) {
+           const appHtml = appHtmls.join('');
+           const startIdxArea = portalContent.search(`<[a-zA-Z]+.*id=["']${appAreaId}["']`);
+           const beginAreaContent = portalContent.indexOf('>', startIdxArea) + 1;
+           portalContent = [portalContent.slice(0, beginAreaContent), appHtml, portalContent.slice(beginAreaContent)].join('');
+       }
+    });
+
+    return portalContent;
+}
+
 const renderToString = (template: string, templateMustExist: boolean, model: any, fallback: () => string, res: Response, logger: MashroomLogger): Promise<string> => {
     return new Promise((resolve) => {
         res.render(template, model, (error, html) => {
@@ -79,3 +129,5 @@ const renderToString = (template: string, templateMustExist: boolean, model: any
         });
     })
 }
+
+const getSafePluginName = (pluginName: string) => pluginName.toLowerCase().replace(/ /g, '-');