+++ packages/plugin-packages/mashroom-portal/test/backend/controllers/PortalPageRenderController.test.ts
@@ -328,7 +328,7 @@ describe('PortalPageRenderController', () => {
                     expect(model.siteBasePath).toBe('/portal/web');
                     expect(model.resourcesBasePath).toBe('/portal/web/___/theme/my-theme');
                     expect(model.apiBasePath).toBe('/portal/web/___/api');
-                    expect(model.portalLayout).toContain('<div class="row"><div id="app-area1"></div><div id="app-area2"></div></div>');
+                    expect(model.pageContent).toContain('<div class="row"><div /><div /><div /><div id="app-area1"></div><div id="app-area2"></div></div>');
 
                 }